Deprecates pantheon-cache in favor of pantheon cache
for more information, see https://pantheon.io/docs/start-states/wordpress#20220816
