+++ wp-content/mu-plugins/pantheon.php
@@ -3,7 +3,7 @@
  * Plugin Name: Pantheon
  * Plugin URI: https://pantheon.io/
  * Description: Building on Pantheon's and WordPress's strengths, together.
- * Version: 0.2
+ * Version: 0.3
  * Author: Pantheon
  * Author URI: https://pantheon.io/
  *
@@ -22,14 +22,17 @@ if ( isset( $_ENV['PANTHEON_ENVIRONMENT'] ) ) {
 	if ( 'dev' === $_ENV['PANTHEON_ENVIRONMENT'] && function_exists( 'wp_is_writable' ) ) {
 		require_once 'pantheon/pantheon-plugin-install-notice.php';
 	}
+    if ( defined( 'WP_CLI' ) && WP_CLI ) {
+        require_once 'pantheon/cli.php';
+    }
 	if ( ! defined( 'FS_METHOD' ) ) {
 		/**
-		 * When this constant is not set, WordPress writes and then deletes a 
+		 * When this constant is not set, WordPress writes and then deletes a
 		 * temporary file to determine if it has direct access to the filesystem,
-		 * which we already know to be the case.  This multiplies filesystem 
+		 * which we already know to be the case.  This multiplies filesystem
 		 * operations and can degrade performance of the filesystem as a whole in
-		 * the case of large sites that do a lot of filesystem operations.  
-		 * Setting this constant to 'direct' tells WordPress to assume it has 
+		 * the case of large sites that do a lot of filesystem operations.
+		 * Setting this constant to 'direct' tells WordPress to assume it has
 		 * direct access and skip creating the extra temporary file.
 		 */
 		define( 'FS_METHOD', 'direct' );
